Moved methods to Controller for accessibility
