+++ framework/src/main/java/fi/otavanopisto/pyramus/koski/exception/NoLatestStudentException.java
@@ -1,0 +1,5 @@
+package fi.otavanopisto.pyramus.koski.exception;
+
+public class NoLatestStudentException extends Exception {
+  private static final long serialVersionUID = 4626848627742926057L;
+}
